+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
+use App\Models\Company;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
@@ -14,7 +16,20 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::all();
+        $users = User::all();
+        return view("users.index", compact("users"));
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
+        $this->authorize("create", User::class);
+        $companies = Company::all();
+        return view("users.create", compact("companies"));
     }
 
     /**
@@ -25,7 +40,15 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        return User::create($request->all());
+        $this->authorize("create", User::class);
+        return User::create([
+            "firstname" => $request->firstname,
+            "lastname" => $request->lastname,
+            "email" => $request->email,
+            "company" => $request->company,
+            "role" => $request->role,
+            "password" => Hash::make($request->password),
+        ]);
     }
 
     /**
@@ -40,6 +63,19 @@ class UserController extends Controller
     }
 
     /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
+        $user = User::findOrFail($id);
+        $companies = Company::all();
+        return view("users.edit", compact(["user", "companies"]));
+    }
+
+    /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
@@ -48,8 +84,13 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $filtered = array_filter($request->all());
+
+        if (in_array("password", $filtered)) {
+            $filtered["password"] = Hash::make($filtered["password"]);
+        }
         $user = User::findOrFail($id);
-        $user->update($request->all());
+        $user->update($filtered);
 
         return $user;
     }
@@ -62,9 +103,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize("delete", User::class);
         $user = User::findOrFail($id);
         $user->delete();
-
+        // TODO: Replace response with toast notification.
         return 204;
     }
 }
